backend/requirement.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.7.4
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
distlib==0.3.9
distro==1.9.0
ecdsa==0.19.0
fastapi==0.115.2
filelock==3.16.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
Jinja2==3.1.4
jiter==0.6.1
MarkupSafe==3.0.1
openai==1.42.0
platformdirs==4.3.6
pyasn1==0.6.1
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.12
requests==2.32.3
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.40.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.31.1
virtualenv==20.26.6
